프:1.사용자 기반 출력 2.전체ui 및 연동, 리뷰작성 폼 작성 및 기능연동 완료 백:객체 변경및 api 수정

+++ backend/flim/src/main/java/com/example/flim/dto/MovieDetailDTO.java
@@ -8,7 +8,7 @@ public class MovieDetailDTO {
  private int review_idx;
  private int user_idx;
  private String nickname;
- private int review_rating;
+ private float review_rating;
  private String review_comment;
  private String write_date;
  private String update_date;
